+++ example/lib/home_page.dart
@@ -6,11 +6,7 @@ class HomePage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return const Scaffold(
-      body: SafeArea(
-        child: Center(
-          child: Text('Home ✅'),
-        ),
-      ),
+      body: SafeArea(child: Center(child: Text('Home ✅'))),
     );
   }
 }